converted test_filter_tree to use GoogleTest
